+++ README.rst
@@ -1,23 +1,23 @@
-========================================
-Advanced Sunbeam OpenStack Documentation
-========================================
+===============================
+Sunbeam OpenStack Documentation
+===============================
 
 Tutorials
 #########
 
-`Writing an OpenStack API charm with ASO <writing-OS-API-charm.rst>`_.
+`Writing an OpenStack API charm with Sunbeam <doc/writing-OS-API-charm.rst>`_.
 
 How-Tos
 #######
 
 
-`How-To write a pebble handler  <howto-pebble-handler.rst>`_.
+`How-To write a pebble handler  <doc/howto-pebble-handler.rst>`_.
 
-`How-To write a relation handler  <howto-relation-handler.rst>`_.
+`How-To write a relation handler  <doc/howto-relation-handler.rst>`_.
 
-`How-To write a charm context  <howto-config-context.rst>`_.
+`How-To write a charm context  <doc/howto-config-context.rst>`_.
 
-`How-To expose services outside of K8S <howto-expose-services.rst>`_.
+`How-To expose services outside of K8S <doc/howto-expose-services.rst>`_.
 
 
 Reference
@@ -29,4 +29,4 @@ Concepts
 ########
 
 
-`ASO Concepts <concepts.rst>`_.
+`Sunbeam Concepts <doc/concepts.rst>`_.
