+++ requirements.txt
@@ -3,7 +3,6 @@ jinja2
 kubernetes
 ops
 python-keystoneclient
-git+https://opendev.org/openstack/charm-ops-openstack#egg=ops_openstack
 git+https://github.com/openstack/charm-ops-interface-ceph-client#egg=interface_ceph_client
 lightkube
 lightkube-models
