+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = advanced_sunbeam_openstack
+name = ops_sunbeam
 summary = Charm lib for OpenStack Charms using operator framework
 version = 0.0.1.dev1
 description-file =
