+++ shared_code/sunbeam_charm/{{cookiecutter.service_name}}/config.yaml
