+++ shared_code/sunbeam_charm/{{cookiecutter.service_name}}/metadata.yaml
