+++ shared_code/sunbeam_charm/{{cookiecutter.service_name}}/requirements.txt
@@ -1,7 +1,6 @@
 ops
 jinja2
-git+https://opendev.org/openstack/charm-ops-openstack#egg=ops_openstack
-git+https://github.com/openstack-charmers/advanced-sunbeam-openstack#egg=advanced_sunbeam_openstack
+git+https://github.com/openstack-charmers/advanced-sunbeam-openstack#egg=ops_sunbeam
 lightkube
 # These are only needeed if the charm relates to ceph
 git+https://github.com/openstack/charm-ops-interface-ceph-client#egg=interface_ceph_client
