+++ shared_code/sunbeam_charm/{{cookiecutter.service_name}}/src/templates/ceph.conf.j2
