+++ shared_code/sunbeam_charm/{{cookiecutter.service_name}}/src/templates/parts/database-connection
