+++ sunbeam-charm-init.sh
@@ -2,4 +2,4 @@
 
 [ -e .tox/cookie/bin/activate ] || tox -e cookie
 source .tox/cookie/bin/activate
-shared_code/aso-charm-init.py $@
+shared_code/sunbeam-charm-init.py $@
