+++ tox.ini
@@ -18,7 +18,7 @@ skip_missing_interpreters = False
 requires = pip < 20.3
            virtualenv < 20.0
 # NOTE: https://wiki.canonical.com/engineering/OpenStack/InstallLatestToxOnOsci
-minversion = 3.2.0
+minversion = 3.18.0
 
 [testenv]
 setenv = VIRTUAL_ENV={envdir}
@@ -71,7 +71,22 @@ deps = -r{toxinidir}/requirements.txt
 basepython = python3
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
-commands = flake8 {posargs} src unit_tests tests advanced_sunbeam_openstack --exclude unit_tests/lib
+commands = flake8 {posargs} src unit_tests tests ops_sunbeam --exclude unit_tests/lib
+
+[testenv:cover]
+basepython = python3
+deps = -r{toxinidir}/requirements.txt
+       -r{toxinidir}/test-requirements.txt
+setenv =
+    {[testenv]setenv}
+    PYTHON=coverage run
+commands =
+    coverage erase
+    stestr run --slowest {posargs}
+    coverage combine
+    coverage html -d cover
+    coverage xml -o cover/coverage.xml
+    coverage report
 
 [coverage:run]
 branch = True
@@ -81,7 +96,6 @@ source =
     .
 omit =
     .tox/*
-    */charmhelpers/*
     unit_tests/*
 
 [testenv:venv]
